Fixed corner case. Created project factories test utilities
